Fix output scores dtype

+++ merlin/systems/dag/ops/softmax_sampling.py
@@ -102,7 +102,7 @@ class SoftmaxSampling(InferenceOperator):
         return Schema(
             [
                 ColumnSchema("ordered_ids", dtype=np.int32, dims=(None, 1)),
-                ColumnSchema("ordered_scores", dtype=np.int32, dims=(None, 1)),
+                ColumnSchema("ordered_scores", dtype=np.float32, dims=(None, 1)),
             ]
         )
